visualizing of mistakes

+++ main.cpp
@@ -1,6 +1,7 @@
 #include <tuple>
 #include "csv_parser.hpp"
 #include "my_tuple.hpp"
+#include "hi_user.hpp"
 
 using namespace csv_parser;
 
@@ -9,7 +10,7 @@ int main() {
 
     std::ifstream in("test.csv");
     std::ofstream out("output.txt");
-    CSVParser<int, int, std::string> parser{ in, skip };
+    CSVParser<int, double, std::string> parser{ in, skip };
 
     for (auto& iter : parser) {
         out << iter << '\n';
